set up new redux structure
