+++ src/actions/burgerBurger.js
@@ -1,0 +1,17 @@
+/* eslint-disable import/prefer-default-export */
+import { makeSynActionCreator } from 'utils';
+
+export const burgerActionNames = {
+  addIngredient: 'ADD_INGREDIENT',
+  removeIngredient: 'REMOVE_INGREDIENT',
+};
+
+export const addIngredientAction = makeSynActionCreator(
+  burgerActionNames.addIngredient,
+  'ingredient',
+);
+
+export const removeIngredientAction = makeSynActionCreator(
+  burgerActionNames.removeIngredient,
+  'ingredient',
+);
